--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Rvv_test.java
@@ -1,4 +1,0 @@
-package org.firstinspires.ftc.teamcode;
-
-public class Rvv_test {
-}
