--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/am_test.java
@@ -1,5 +1,0 @@
-package org.firstinspires.ftc.teamcode;
-
-public class am_test {
-
-}
