+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/otherTuning/Drawing.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.otherTuning;
 
 import com.acmerobotics.dashboard.canvas.Canvas;
 import com.acmerobotics.roadrunner.Pose2d;
