+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/tuning/LocalizationTest.java
@@ -8,9 +8,9 @@ import com.acmerobotics.roadrunner.PoseVelocity2d;
 import com.acmerobotics.roadrunner.Vector2d;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
-import org.firstinspires.ftc.teamcode.Drawing;
-import org.firstinspires.ftc.teamcode.MecanumDrive;
-import org.firstinspires.ftc.teamcode.TankDrive;
+import org.firstinspires.ftc.teamcode.otherTuning.Drawing;
+import org.firstinspires.ftc.teamcode.otherTuning.MecanumDrive;
+import org.firstinspires.ftc.teamcode.otherTuning.TankDrive;
 
 public class LocalizationTest extends LinearOpMode {
     @Override
